+++ srcjs/components/oauth.js
@@ -12,6 +12,9 @@ let oauthProviders = [];
 
 Shiny.addCustomMessageHandler('fireblaze-set-oauth-provider', (msg) => {
 	oauthProviders[msg.id] = new OAuthProvider(msg.provider);
+
+	if(Object.entries(msg.opts).length == 0)
+		oauthProviders[msg.id].setCustomParameters(msg.opts)
 });
 
 Shiny.addCustomMessageHandler('fireblaze-oauth-sign-in-popup', (msg) => {