+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - Update firebase to `9.6.7` to fix 
 [#23](https://github.com/JohnCoene/firebase/issues/23)
+- Allow passing parameters to OAuth providers fixes [#33](https://github.com/JohnCoene/firebase/issues/33)
+- Fix instructions to setup Microsoft
 
 # firebase 1.0.1